+++ lib/rails_uuid_pk.rb
@@ -6,6 +6,7 @@ require "rails_uuid_pk/type"
 require "rails_uuid_pk/uuid_adapter_extension"
 require "rails_uuid_pk/sqlite3_adapter_extension"
 require "rails_uuid_pk/mysql2_adapter_extension"
+require "rails_uuid_pk/trilogy_adapter_extension"
 require "rails_uuid_pk/railtie"
 
 # Load generators